Implements feature #155 - Read only attributes.
This commit implements the read only support for attribute editing in the
backend on global level.
The widget has to honor the setting to make it visible to the user in the
browser.
There are issues with (at least) tinyMCE and "chosen"-enabled selects
which appear to be editable in the browser but are in fact not. The MetaModel
will refuse to save such changed input and revert the content to the previous
state thanks to DC_General for those corner cases. Therefore this issue only
persists optically and has to be resolved in the various browsers and/or via
JavaScripts.