MetaModel::fetchRows() now honors to honor attributeOnly (see #158).
The behaviour of MetaModel::fetchRows() has been altered to only
return the requested attributes, if any restriction is being applied,
or all attributes if no restriction is being applied (as the behaviour
was before).

NOTE: This breaks backwards compatibility as simple columns will not
be included in result values anymore, if not being requested via
rendersettings or the like.