Update SimpleLookup.php
Hi, I had much problems to set properties for a Filter which I had generated in the Template. The main difficulty was that I use a custom sql in the list because it's complex. but in the filter I couldn't build the feature onlyused and onlypossible. That is because the attribute is only referenced and part of an other table/metamodel. So with this custom array I can modify the important properties.

What do you think about this solution =)?

 /* $arrCustom array with settings which makes the function more flexible to use in template
            the following keys can be set: objAttribute, forceToUseArrIds, strParamName, 
            [Widget > eval: includeBlankOption, onlyused, onlypossible, template] (only eval for now)
        */