Bugfix for the creating of a new item as varbase.
Add a new event which checks the current item. If the MetaModels has variant support and the current item has a empty vargroup it must be a varbase. So the varbase is set to '1'. This should fix a bug where the base attributes weren't displayed for a new base item.