Fix translated reference "hole" handling
The implementation we have, i.e. in translated checkbox, does not save
"empty" values to the database but creates them on demand in
getTranslatedDataFor(), however this interferes with our optimization to
first check the contained items, then update the existing ones and
insert the new ones later on.
The current approach tried to update non existing values in the database
which always succeeded with "0 rows affected" and therefore never could
save non empty values anymore.