Fix #1318 - allow empty values in "show"
Empty values must be allowed in "show" when "hideEmptyValues" is active.

The empty value checks have now been refactored into a static helper
class as the detection of empty values is not within the attribute
classes and therefore is somewhat bogus.
We should refactor this for 3.0 to have attributes determine their empty
values on their own.