Change the handling of tags in the custom sql filter settging
In the old version, each insert tag, with only one part, was
interpreted by contao. In the current version, instead of running
the "replaceInsertTag" function, the tag was given back
as a plain string. This patch change the behavior that even
a simple insert tag will be interpreted by the contao
"replaceInsertTag" function.